pyproject.toml
[tool.poetry]
name = "l00192944_artefact"
version = "0.1.0"
description = "L00162944 Project Artefact for CSD LYIT"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
celery = "^5.1.2"
django-celery-beat = "^2.2.1"
idna = "^3.2"
cryptography = "^3.4.7"
pyOpenSSL = "^20.0.1"
requests = "^2.26.0"
dnspython = "^2.1.0"
python3-nmap = "^1.5.0"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.6b0", allow-prereleases = true}
django-extensions = "^3.1.3"
pydotplus = "^2.0.2"
graphviz = "^0.17"
pre-commit = "^2.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
